Adds a quick about page and some minor tweaks - media queries, some reorganization to categorize components by pages.
