+++ src/App.js
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
 import './App.css';
-import Home from './Components/Home/Home';
-import Explore from './Components/Explore/Explore';
+import Home from './Components/HomePg/Home/Home';
+import Explore from './Components/ExploreTrackedPgs/Explore/Explore';
 import Error from './Components/Error/Error';
-import BirdDetails from './Components/BirdDetails/BirdDetails';
-import OwlingTips from './Components/OwlingTips/OwlingTips';
+import About from './Components/About/About';
+import BirdDetails from './Components/BirdDetailsPg/BirdDetails/BirdDetails';
+import OwlingTips from './Components/OwlingTipsPg/OwlingTips/OwlingTips';
 import { Switch, Route } from 'react-router-dom';
 
 const App = () => {
@@ -35,6 +36,9 @@ const App = () => {
       </Route>
       <Route exact path='/owling-tips'>
         <OwlingTips />
+      </Route>
+      <Route exact path='/about'>
+        <About />
       </Route>
       <Route exact path='/error'>
         <Error />
