+++ src/Components/BirdDetailsPg/BirdDetails/BirdDetails.css
