+++ src/Components/BirdDetailsPg/CollapsibleInfo/CollapsibleInfo.css
