+++ src/Components/BirdDetailsPg/CollapsibleInfo/CollapsibleInfo.js
