+++ src/Components/BirdDetailsPg/HighLevelnfo/HighLevelInfo.css
