--- src/Components/Map/Map.css
@@ -1,8 +1,0 @@
-.leaflet-container {
-    height: 80%;
-    width: 100%;
-    align-self: center;
-    border-radius: 15px;
-    grid-row: 2;
-    grid-column: 3;
-}
